Refactor logic for changing the animation
Previously we had a bunch of awkward calculations to avoid resetting
the animation when nothing had changed (and thus resetting to the
first frame of the animation). I've refactored to separate the logic
of which state it's currently in (a low-cost thing that we can
calculate on each frame) from the act of changing the animation (which
we only do when necessary).